requirements.txt
aider-chat
GitPython 